kfc-backend/requirements.txt
boto3>=1.26.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.0
pydantic>=2.0.0
python-dateutil>=2.8.2
ulid-py>=1.1.0
